Make PSR-7 an OPTIONAL dependency
File filters are a minority use case of the component, and adding a
dependency on psr/http-message and zendframework/zend-diactoros is
overkill if you are not using them. As such, I have made
psr/http-message-implementation a suggested dependency, and
zendframework/zend-diactoros a development dependency.

I have also allowed either the 1.X or 2.X series of Diactoros, as they
both fulfill PSR-7.

Finally, to allow tests to run against 5.6 and 7.0, I have added
Diactoros as a legacy dependency (so that the 1.X version can be
installed), as well as zend-crypt (but only against 5.6; so that
paragonie/random_compat can receive a version compatible with PHP 5).