Refactor: extract method filterPsr7UploadedFile
This patch extracts the logic specific to filtering a PSR-7
`UploadedFileInterface` instance to its own method,
`filterPsr7UploadedFile()`. `filter()` now detects
`UploadedFileInterface` instances early in the method, and, when found,
returns the return value of the new method. This keeps the logic within
`filter()` simpler, and ensures all logic related to filtering PSR-7
artifacts is in one place.