Compiling the snappy extension is breaking CI
It's not worth fixing the build of the snappy extension given that it's deprecated for removal ASAP.

The failed build is causing `extension not found` errors which in turn is causing `phpcs` to fail.

The build has been failing with these errors for a while so tests have just been skipped anyway.

Signed-off-by: George Steel <george@net-glue.co.uk>