Refactor Lower Case and Upper Case Filters
- Removes `AbstractUnicode` from the inheritance tree and deletes it
- Clean up `File\LowerCase`, `File\UpperCase`, `StringToLower`, `StringToUpper` and `UpperCaseWords` so that they all implement `FilterInterface` directly without any inheritance.

All the filters support an `encoding` option which is re-implemented in each class. Boilerplate is reduced by adding a new internal class `EncodingOption` for consistent handling of encoding options…

Extracted some duplicated file handling code into an internal class `FilterFileContents` that applies a filter to the contents of a file.

This removes the need for `File\LowerCase` to extend `StringToLower` for example.

Signed-off-by: George Steel <george@net-glue.co.uk>