Convert blockquotes to admonitions/call-outs
Admonitions are used to highlight important information in the documentation.
They are used to draw attention to specific information, such as warnings, tips, or notes.
Admonitions are often used to provide additional context or to explain the implications of a particular action.

## Basic Usage

Start the admonition with one of the types in capital letters followed by a colon.
The type corresponds with a specific style.
The type should be followed by the content of the admonition.

```markdown
NOTE: An example of a note.
```

Signed-off-by: Frank Brückner <dev@froschdesignstudio.de>